Гетеры и сетеры переименовал. Классы разделил. show_product возвращает список
