+++ main.py
@@ -1,4 +1,3 @@
-from classes import open_json
-from function import create_instance
+from function import create_instance, open_json
 
 print(create_instance(open_json())[0][0].show_product)
